+++ templates/tiny-todos/src/client/index.js
@@ -1,7 +1,8 @@
 /* eslint-env browser */
 
 import React from 'react';
-import { render } from 'react-dom';
+import ReactDOM from 'react-dom';
+import { AppContainer } from 'react-hot-loader';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 
@@ -10,6 +11,22 @@ import reducer from './reducer';
 
 const store = createStore(reducer);
 
-render(
-  React.createElement(Provider, { store }, React.createElement(App)),
-  document.getElementById('app'));
+const render = (Component) => {
+  ReactDOM.render(
+    React.createElement(AppContainer, {},
+      React.createElement(Provider, { store }, React.createElement(Component))),
+    document.getElementById('app'));
+};
+
+render(App);
+
+if (module.hot) {
+  module.hot.accept('./App', () => {
+    const NewApp = require('./App').default;
+    render(NewApp);
+  });
+  module.hot.accept('./reducers', () => {
+    const newReducer = require('./reducer').default;
+    store.replaceReducer(newReducer);
+  });
+}